requirements.txt
flask==2.3.3
flask-wtf==1.1.1
python-dotenv==1.0.0
pillow==10.0.0
requests==2.31.0
